+++ src/contracts/common/model/attribute-values-type.ts
@@ -3,7 +3,7 @@ import {z} from "zod";
 export const ZAttributeValues = z.record(
     z.string(),
     z.unknown(),
-).optional();
+).optional().default({});
 
 /* Custom data */
 export type AttributeValues = z.infer<typeof ZAttributeValues>;